Migrate utils functions
