+++ lib/configuration/hooks/jsonapi/helpers/response.js
@@ -7,6 +7,7 @@
 // Public node modules.
 const _ = require('lodash');
 const JSONAPISerializer = require('jsonapi-serializer');
+const utils = require('../utils/utils');
 
 /**
  * JSON API helper
@@ -21,8 +22,8 @@ module.exports = {
    */
 
   set: function (ctx, matchedRoute, actionRoute) {
-    const object = this.getObject(matchedRoute);
-    const type = this.getType(ctx, actionRoute.controller);
+    const object = utils.getObject(matchedRoute);
+    const type = utils.getType(ctx, actionRoute.controller);
 
     // Fetch a relationship that does not exist
     // Reject related request with `include` parameter
@@ -59,7 +60,7 @@ module.exports = {
       }
     };
 
-    const PK = this.getPK(type);
+    const PK = utils.getPK(type);
 
     if (_.isArray(value) && !_.isEmpty(value)) {
       // Array
@@ -124,7 +125,7 @@ module.exports = {
           case 'manyToOne':
             // Object
             toSerialize[relation.alias] = {
-              ref: self.getPK(relation.model),
+              ref: utils.getPK(relation.model),
               attributes: _.keys(strapi.models[type].attributes),
               relationshipLinks: {
                 self: function (record) {
@@ -145,7 +146,7 @@ module.exports = {
           case 'manyToMany':
             // Array
             toSerialize[relation.alias] = {
-              ref: self.getPK(relation.collection),
+              ref: utils.getPK(relation.collection),
               typeForAttribute: relation.collection,
               attributes: _.keys(strapi.models[type].attributes),
               relationshipLinks: {
@@ -210,66 +211,5 @@ module.exports = {
       default:
         return 'collection';
     }
-  },
-
-  /**
-   * Find data object
-   */
-
-  getObject: function (matchedRoute) {
-    // TODO:
-    // - Improve way to detect collection/ressource/relationships/related
-    switch (_.size(matchedRoute.regexp.keys)) {
-      case 0:
-        return 'collection';
-      case 1:
-        return 'ressource';
-      case 2:
-        return (matchedRoute.path.indexOf('relationships') !== -1) ? 'relationships' : 'related';
-      default:
-        return 'collection';
-    }
-  },
-
-  /**
-   * Find data type
-   */
-
-  getType: function (ctx, supposedType) {
-    // TODO:
-    // - Parse the URL and try to extract useful information to find the type
-
-    // Relationships or related ressource
-    if (strapi.models.hasOwnProperty(supposedType.toLowerCase()) && ctx.params.hasOwnProperty('relation')) {
-      return _.first(_.reject(_.map(strapi.models[supposedType.toLowerCase()].associations, function (relation) {
-        return (ctx.params.hasOwnProperty('relation') && ctx.params.relation === relation.alias) ? relation.model || relation.collection : undefined;
-      }), _.isUndefined));
-    } else if (strapi.models.hasOwnProperty(supposedType.toLowerCase())) {
-      return supposedType.toLowerCase();
-    }
-
-    return undefined;
-  },
-
-  /**
-   * Find primary key
-   */
-
-  getPK: function (type) {
-    if (!strapi.models.hasOwnProperty(type)) {
-      return null;
-    }
-
-    const PK = _.findKey(strapi.models[type].attributes, {primaryKey: true});
-
-    if (!_.isUndefined(PK)) {
-      return PK;
-    } else if (strapi.models[type].attributes.hasOwnProperty('id')) {
-      return 'id';
-    } else if (strapi.models[type].attributes.hasOwnProperty('uuid')) {
-      return 'uuid';
-    }
-
-    return null;
   }
 };
